refactor: hoist my games related stuff to global context
All components share the same `myGames` state, so when any component
modifies it, the state is propagated for everyone.

👍🌍
