+++ app/src/MyGames/index.js
@@ -1,7 +1,7 @@
 import * as React from "react";
 
 import GameList from "../components/GameList";
-import { useMyGames } from "../hooks";
+import { useMyGames } from "../myCollection";
 
 const MyGames = () => {
   const { myGames } = useMyGames();
