+++ app/src/components/GameList/index.js
@@ -3,48 +3,48 @@ import * as React from "react";
 import { FaMinusSquare, FaPlusSquare } from "react-icons/fa";
 import { Link } from "react-router-dom";
 
-import { useMyGames } from "../../hooks";
+import { useMyGames } from "../../myCollection";
 
 import styles from "./styles.module.scss";
 
-const GameList = ({ games }) => {
+const GameList = ({ games }) => (
+  <>
+    <ul className={styles.grid}>
+      {games.map((game) => (
+        <li className={styles.card} key={game.id}>
+          <Card {...{ game }} />
+        </li>
+      ))}
+    </ul>
+  </>
+);
+
+const Card = ({ game }) => {
   const { addGame, deleteGame, isInMyGames } = useMyGames();
 
   return (
     <>
-      <ul className={styles.grid}>
-        {games.map((game) => (
-          <li className={styles.card} key={game.id}>
-            <Card {...{ game, addGame, deleteGame, isInMyGames }} />
-          </li>
-        ))}
-      </ul>
+      <div className={styles.wrapper}>
+        <div key={game.id} className={`${styles.box} ${styles.dropshadow}`}>
+          <header>
+            <Link to={`/games/${game.id}`}>{game.name}</Link>
+          </header>
+          {game.image_url ? (
+            <img
+              src={game.image_url}
+              alt={game.name}
+              className={styles.cardthumb}
+            />
+          ) : null}
+          {isInMyGames(game.id) ? (
+            <FaMinusSquare onClick={() => deleteGame(game)} />
+          ) : (
+            <FaPlusSquare onClick={() => addGame(game)} />
+          )}
+        </div>
+      </div>
     </>
   );
 };
-
-const Card = ({ game, addGame, deleteGame, isInMyGames }) => (
-  <>
-    <div className={styles.wrapper}>
-      <div key={game.id} className={`${styles.box} ${styles.dropshadow}`}>
-        <header>
-          <Link to={`/games/${game.id}`}>{game.name}</Link>
-        </header>
-        {game.image_url ? (
-          <img
-            src={game.image_url}
-            alt={game.name}
-            className={styles.cardthumb}
-          />
-        ) : null}
-        {isInMyGames(game.id) ? (
-          <FaMinusSquare onClick={() => deleteGame(game)} />
-        ) : (
-          <FaPlusSquare onClick={() => addGame(game)} />
-        )}
-      </div>
-    </div>
-  </>
-);
 
 export default GameList;
